Add windows compatibility. (#933)
* Add windows compatibility.

* Fix CI

* Fix benchmark for ci

* Exclude bench from test.

* Remove baseline for vcpkg

* Add baseline back for vcpkg.

* Update to latest baseline.

* Make windows CI its own config.

* Fix windows build.

* Reduce windows build scope.

* Breakout CI builds and try to improve windows CI.

* Fix windows ninja compile flags.

* Bette rheader macro inclusion for sockets and remove a few warnings.

* Prevent linter from breaking on reasonable standard macro definition.

+++ benchmark/safe_time_queue.cpp
@@ -32,7 +32,7 @@ BM_SafeTimeQueue_Push(benchmark::State& state)
 
         {
             std::lock_guard __(tq);
-            tq.Push(items_count, 20);
+            tq.Push(static_cast<int>(items_count), 20);
         }
     }
 
@@ -44,7 +44,7 @@ BM_SafeTimeQueue_Pop(benchmark::State& state)
 {
     quicr::SafeTimeQueue<int> tq(state.range(0), 1, service, kIterations);
     for (size_t i = 0; i < kIterations; ++i) {
-        tq.Push(i, 10);
+        tq.Push(static_cast<int>(i), 10);
     }
 
     int64_t items_count = 0;
@@ -65,7 +65,7 @@ BM_SafeTimeQueue_Front(benchmark::State& state)
 {
     quicr::SafeTimeQueue<int> tq(state.range(0), 1, service, kIterations);
     for (size_t i = 0; i < kIterations; ++i) {
-        tq.Push(i, 15);
+        tq.Push(static_cast<int>(i), 15);
     }
 
     int64_t items_count = 0;
@@ -88,7 +88,7 @@ BM_SafeTimeQueue_PopFront(benchmark::State& state)
 {
     quicr::SafeTimeQueue<int> tq(state.range(0), 1, service, kIterations);
     for (size_t i = 0; i < kIterations; ++i) {
-        tq.Push(i, 15);
+        tq.Push(static_cast<int>(i), 15);
     }
 
     int64_t items_count = 0;
@@ -111,7 +111,7 @@ BM_SafeTimeQueue_Size(benchmark::State& state)
 {
     quicr::SafeTimeQueue<int> tq(300, 1, service, kIterations);
     for (size_t i = 0; i < kIterations; ++i) {
-        tq.Push(i, 10);
+        tq.Push(static_cast<int>(i), 10);
     }
 
     for (auto _ : state) {
@@ -130,7 +130,7 @@ BM_SafeTimeQueue_Empty(benchmark::State& state)
 {
     quicr::SafeTimeQueue<int> tq(300, 1, service, kIterations);
     for (size_t i = 0; i < kIterations; ++i) {
-        tq.Push(i, 10);
+        tq.Push(static_cast<int>(i), 10);
     }
 
     for (auto _ : state) {
@@ -153,7 +153,7 @@ BM_SafeTimeQueue_PushAndPopLoaded(benchmark::State& state)
 
     for (auto _ : state) {
         ++items_count;
-        tq.Push(items_count, 1000);
+        tq.Push(static_cast<int>(items_count), 1000);
 
         // Simulate load by not popping all items
         if (items_count % 100 == 0) {
@@ -184,7 +184,7 @@ BM_SafeTimeQueue_PushAndPop_Interval_1ms(benchmark::State& state)
         {
             std::lock_guard __(tq);
 
-            tq.Push(items_count, 1000);
+            tq.Push(static_cast<int>(items_count), 1000);
 
             auto elem = tq.Front();
             tq.Pop();
@@ -210,7 +210,7 @@ BM_SafeTimeQueue_PushAndPop_Interval_125ms(benchmark::State& state)
         {
             std::lock_guard __(tq);
 
-            tq.Push(items_count, 1000);
+            tq.Push(static_cast<int>(items_count), 1000);
 
             auto elem = tq.Front();
             tq.Pop();
